Updated build.sh to be fancy and how

+++ build.sh
@@ -5,6 +5,8 @@ cd ../../
 # Generate new Data file
 java com.hut.DataGenerator outfile.txt 10
 
+
+: ${JAVA_HOME:?"Need to set JAVA_HOME, on lab machines this is /usr/lib/jvm/java-1.7.0"}
 rm *.so
 javah -jni com.hut.InsertionSort
 cp com_hut_InsertionSort.h com/hut/
